requirements.txt
gymnasium>=0.29.0
numpy>=1.23.0
matplotlib>=3.6.0
pandas>=1.5.0
seaborn>=0.12.0
pyyaml>=6.0
stable-baselines3==2.0.0
torch>=1.12.0
tensorboard>=2.12.0
optuna>=3.1.0
tqdm>=4.65.0